Python/cloudbeat-pytest-selenium/requirements.txt
# cloudbeat-common
# cloudbeat-pytest
# cloudbeat-selenium
cloudbeat-selenium-test
cloudbeat-pytest-test
cloudbeat-common-test
cloudbeat-playwright-test
attrs
certifi
cffi
charset-normalizer
colorama
et_xmlfile
execnet
h11
idna
iniconfig
Jinja2
MarkupSafe
openpyxl
outcome
packaging
pluggy
pycparser
PySocks
pytest
pytest-html
pytest-metadata
pytest-xdist
python-dotenv
webdriver-manager
requests
selenium
setuptools
sniffio
sortedcontainers
trio
trio-websocket
typing_extensions
urllib3
websocket-client
websockets
wsproto
uuid